Xc-6514 - Additional error output for expServer.js
This bug exposed two problems with error reporting for the behavior
of the back end:

1) The expServer does not log the command it sends to the back end
of the GUI installer, making it difficult to determine if bad input
is causing installer problems.

2) The expServer collects stdout and stderr information from the
command it sends to the back end of the installer and puts it
in the errorLog.  The contents of errorLog are only reported to
the web gui by the sendStatusArray routine when the backend
detects a failure (when status != Status.Ok).  In this case,
the command failed immediately, without ever updating status.
The /checkStatus routine completed successfully and never
returned an error.

This change to expServer.js does two things:

1) Upon error, it logs the command sent to the back end.

2) It also logs the errorLog collected from that command.

Change-Id: Iecbca719e7d1a7a862196f16554378cff43150f2