Xc-7407 - identifying Xcalar administrators when using Active Directory
Changes in expLogin.js collect the memberOf attribute instead of
employeeType when Active Directory is used, which should return a list
of the groups to which the AD user belongs.  If one of the groups begins
with 'Xce Admin', the user is designated to be a Xcalar administrator
user.

Also, rather than use the LDAP filter input field to separate xcalar
users from non-users, the code searches for membership in a group
beginning with 'Xce User'. If this is found, isADUser is set to true
in the UserInformation.  If isADUser is false once all LDAP records
are searched, then the login is rejected.

Change-Id: I9897ee2de9de40ae0e7596badbe5bc21e040a893