Configurable User and Admin Group names for Active Directory Support
Previously, the names "Xce User" and "Xce Admin" were hard wired for
detection of Xcalar administrator and user status.

This feature adds the ability to provide two JSON files in the
ldapConfig.json file to set the group names for administrator and
user status to arbitrary values:

{
    "ldap_uri":"ldap://pdc1.int.xcalar.com:389",
    "userDN":"cn=Users,dc=int,dc=xcalar,dc=net",
    "useTLS":"true",
    "searchFilter":"(&(objectclass=user)(userPrincipalName=%username%))",
    "activeDir":"true",
    "serverKeyFile":"/etc/ssl/certs/ca-certificates.crt",
    "adUserGroup":"Schema Admins",
    "adAdminGroup":"Another User Group"
}

If the properties are not in the object, the old defaults will be used.

Change-Id: If3db960efee8409fd92276b4e68c52bf45170849