GUI-8975 - XD stop/start functionality fails for tarball installs
A couple changes to enable communication with Caddy:

http and https ports are no longer hard coded to 80 and 443.  They
can now be initialized with the environment variables XCE_HTTP_PORT
and XCE_HTTPS_PORT.  If those are undefined, they fall back to the
old values.

This also sets the NODE_TLS_REJECT_UNAUTHORIZED environment variable
to 0.  This allows communication with servers with untrusted keys.
We need to be better dealing with that.  There is the rejectUnauthorized
approach:

request({
    "rejectUnauthorized": false,
    "url": domain+"/api/orders/originator/"+id,
    "method": "GET",
    "headers":{
        "X-API-VERSION": 1,
        "X-API-KEY": key
    },
}, function(err, response, body){
    console.log(err);
    console.log(response);
    console.log(body);
});

There is also the method of setting up trusted keystores properly.

Change-Id: I80989339984b1c5f2ba9b72e692dbaf4673aa849