GUI-9636 UI for installer/upgrader/uninstaller
This change makes changes to the GUI and the expServer that enables
an expanded install procedure, as well as upgrade and uninstall
procedures for Xcalar over the 1.2.0 and 1.2.1 installers.

Changes to the GUI include a revised screen layout that loads LDAP
configuration before cluster install, the ability to run pre-config.sh
and verify.sh with the GUI installer, upgrade and uninstall include
the capability for the installer to discover cluster configurations.

Change-Id: I970ab315359b109de77e25603268eed46e6a0aa1