GUI-9649 Add defaultAdmin config to expServer
expServer will now look under
$XLR_INSTALLER_ROOT/config/defaultAdmin.json to see if a default
administrator account is enabled.

This file has to have file permission 600 to ensure that only authorized
users can create and modify the file.

Username is saved in the clear, and a SHA-256 hash of the password is
saved.

If a default admin account is active, you'll be able to login with the
default admin credentials.

Change-Id: I1fd5b309ff4b5c4f248c22fbbbde2decfb379242