GUI-10611 - make optional AD installer fields really optional
This fix changes how AD fields are checked to see if they are filled
in.  The previous check (incorrectly) raised an error if they were not
all populated.  Now they are removed from the JSON data structure
written to disk if they are empty (being optional, they can be empty.)

Change-Id: Ia53730c8e35793195bf00b6b90fb7828e34c3e97