GUI-10845, GUI-10652 Frontend implementation of default admin setup screen in installer
This is part of a change that provides the gui-installer back end with
the ability to create a default admin configuration json file with the
username, e-mail, and encrypted password provided to the installer
interface.

It modifies the GUI to make LDAP configuration optional (by deferring it
to post-install) and allows an optional configuration of a default admin
user.  It also alters the expServer to provide command line arguments
to the back end that create the config file.

Change-Id: I51123c220502f3b98858f3e70c7b92976285024b