[Grunt]: Run watch tasks concurrently; plugin mod
1. The workflow for watch functionality changes.
Will Run grunt watch jobs concurrently - one for files to
livereload and one for files to watch but not livereload.
(Can not have more than one plugin target with livereload
enabled due to port conflict, and can not switch livereload
functionality on the fly, but devs want to be able to
watch certain filetypes but only livereload on a subset
of those filetypes)

2. General requests:

	a. Default to $XLRGUIDIR for project src (current default is cwd)
	b. Default to $XLRDIR for trunk src on trunk builds
	c. Fail if builddir is not descendent of src dir
	d. If watching html, watch for changes in assets/lang/<lang>/htmlTStr.js
		files, and if detected trigger HTML rebuild (Dev request)
	e. Modify initial rsync to handle Ted's Microsoft auth case
	f. 'grunt init' task: to be run when there's a new workspace;
		runs 'npm install' and applies patch of plugin mod (see 3.)

3. The grunt-contrib-watch plugin is modified, so that
less files are NEVER sent to the livereload server.
This allows us to watch css files when less watch with
livereload is requested, and only css files will be sent
to the livereload server, which will force the browser
to refresh without an update. (Dev request)

(Note on 3. This plugin change was done, to accomidate the
usecase of watching less and livereloading the browser
without a refresh.
Why: The alternative, is you would need to run less and css
 in sibling processes
(one running plugin target for less with livereload disabled,
and one running a plugin target for css with livereload enabled),
and catch watch events triggered by one sibling - less -
in the other sibling - the one for css.
Since we are not spawning child processes in individual targets,
(this creates another set of bugs)
the watch plugin will restart automatically after the tasklist
for a watch event concludes.
(So in the less/css case, the user changes a less file,
the process monitoring less starts building css files,
the first css file built triggers the sibling watch process,
while the css re-build is still working,
and the plugin restarts right away since there are no
tasks for css file changes).
The problem: if multiple filetypes are being watched,
the plugin reload takes a few seconds, and in that time,
css edits triggered by the sibling process are missed,
so those filechanges are not detected as watch events
and not sent to the livereload server;
thus the browser does not refresh.)
If the less and css are being monitored in the same process,
this does NOT occur and no watch events are missed since
the plugin will not refresh until the entire css rebuild completes.
However, the plugin behavior is to send the list of all files
changes in the target to the livereload server,
and if less files are sent the browser will do a reload.
THerefore, the plugin was modified to restrict sending
less files to the livereload server, so that less and css
can be monitored in the same target, the the livereload server
can reload the browser without a refresh.

Change-Id: I72e33b203ca723834bafcda6ed84b2af74dd9610