[Grunt] exclude xcalar-infra/ from initial rsync
Jenkins job XDTestSuite, will clone xcalar-infra in to $WORKSPACE.
Therefore, if $WORKSPACE is xcalar-gui, it's going to get cloned
in to xcalar-gui.  Later, grunt will get called.  As a result,
unless xcalar-infra is excluded from the initial rsync, it will
get synced, and eventually tarred, in to the build output.

This change just excludes the directory xcalar-infra/ in initial
rsync.

Change-Id: Ie702d688e4309a1ccc70d43435f526d4f70119d4