V2 License changes for installation
This depends on XCE changes currently in https://gerrit.int.xcalar.com/#/c/7005

This change accomodates the new gzip-compressed, base64-encoded V2
license format.  It checks that the license consists of the base64
character set, but eliminates the length requirement because the
exact number of characters will be variable.  It decodes and
decompresses the license string before passing it to the back end
for installation.

Change-Id: I626ae3f55d045dd4b3a5bed645ee384dc7ad6bfb