Update to the content of GUI installer install Step 2
The question "Were pre-config.sh and verify.sh run on all the
cluster nodes?" in Step 2 of the Xcalar GUI install wizard
proved to cause a significant amount of confusion.  People
answered 'yes' because pre-config.sh was run during a previous
install run (for a failed or now deleted install) and then
were surprised when install options were missing, for example.

The question of whether pre-config.sh was run masks the real
question: what type of installation should the installer perform?
Do users want the installer to perform just non-root tasks, or
do they want it to set up NFS, LDAP, etc...?

This change makes the question in Step 2 about what type of
install is performed.

Change-Id: If6bbb5a0b38ebf4c440ca1efe0441373db13749a