[Gruntfile] watch - rebuild only ts file changed
Currently, if user is running watch --ts and a ts file changes,
will copy that file over to the build and re-reun the BUILD_JS,
however, since other ts src files remain in build, will end up
rebuilding entire ts; this takes a very long time.

Instead, create a temp staging dir for ts.  Copy the ts file there
along with the tsconfig.json file, and run tsc from that dir.
(This is being done on a dir level, as was not finding option to
re-build single ts file while specifying config file as well.)

- At end of the watch process remove the temp staging dir.

Also, if it is the tsconfig.json file itself that is changed,
copy the entire ts/ src in to the temp staging dir, and run tsc
from there so that changes will be reflected.

Also, for watch --js, take out check for rebuilding ts if the js
file is in the ts src folder.  From now on, just copy changed
js directly in to build at same path rel src.

Change-Id: I61bbb16aed37d0993ecd056cebee0dc92caf9e95