OAuth verification for Microsoft Common Endpoint and B2C auth
This change adds a REST endpoint and new auth route to the expServer.
When invoked as a part of MSAL authentication by XD, it takes an
OAuth id token provided by Azure infrastructure and verifies that
Microsoft issued the token.

It does this by issuing web calls that retrieve a JSON list of keys
published by Microsoft that Azure uses to sign tokens.  It then passes
the token and the key through a standard JWT verify call.

Change-Id: Ib153d5949f4c409a34be91814bf6b6fd30d45c27