Change XDTestSuite to use https
With the move to caddy, installs of Xcalar no longer support http.
So, XD testing cannot connect to XD using https, and then have XD
callback to the server using http.  The test server has to respond
to https.  This change adds a certificate and key to that server
and makes the protocol changes necessary in the test handler to
to accept using either http or https on the specified port.

Change-Id: I5eba51cb406b139beb73118dfbf3716f000896dd