GUI-11908 Temporary workaround to default to XcalarMod.Oper
We're no longer returned product == XceOper or XceMod. In fact, we
should not be keying off product name to decide product behavior.

As a temporary workaround, we're setting licenseMode=XcalarMod.Oper
because that turns on all features.

In future, at a feature level, we should be asking the license file if
that particular feature is enabled/disabled instead of depending on
licenseMode.

Change-Id: Idd56b3d5d8b3b4dd92038d5a11619c643c156661