GUI-11671 - New Workbook create takes ~10 seconds
Xcalar performance creating and deleting workbooks is slow because
Jupyter is not starting properly.  Jupyter is not starting properly
because the location of the JupyterNotebooks has been moved to
XLRROOT/JupyterNotebooks as part of the Pr-0081 Workbooks changes
in commit fb69651e3.

This change fixes the problem by having jupyter_notebook_config.py
detect Constants.XcalarRootCompletePath in the XCE_CONFIG file and
look for the JupyterNotebooks subdirectory there.

Change-Id: Icea5cfab10efa8d5b8452f9b9c96db09ff80c939