GUI-11895 IMD fix: Fix key collision on column extraKeyCol
The old update code did a finalizeTable and then a rankOver. This causes the
temporary columns that are created by the rankOver to remain. When this table is
then republished / reupdated we will once again generate the new rankOver
temporary columns, causing a duplicate extraKeyCol (and other columns). The fix
switches the order of finalizeTable and rankOver (this is what publish does as
well). However, with this switch, we can actually clean up the code further
(this review does not do that). We do need to rewrite this extension.

Change-Id: I6bbdf95cff12a0833d0522f29ea2a92d8ee606c6