Xc-11482 - Install Wizard should warn users that the number of nodes they entered is greater than the max nodes their license allows
This change causes uncompressed license data created during license
checking in Step 1 of the Install Wizard to be converted to JSON and
passed back to XD code in InstallerCommon.  It is retrieved via the
InstallerCommon.getLicense() method and compared against the number
of nodes entered by the user in Step 5. If the number of nodes is
greater than that allowed by license, an error message is produced
and the rest of the host information form is not displayed.

Change-Id: I88a40909fc94473fd17262112e32f0ce97be3d9f