expServer tests - auth route tests and fixed login route tests
This change adds tests for the expServer auth route.  The tests
rely on a series of RSA keys generated by openssl (commands to
generate the keys are in comments in authSpec.js).  The RSA keys
are loaded into an array of JWKs (JSON web keys) that functionally
resemble an array provided by Microsoft.  The keys are used to sign
a series of mock JWTs (JSON web tokens) that stand in for OAuth
JWTs created by Microsoft.  The /auth/azureIdToken REST endpoint
is tested various valid and invalid combinations of keys and tokens.

The change also fixes tests of the login route.  The primary problem
here was that ldapConfig.json was getting clobbered by some prior
test in the npm test suite.  The fix for this is to create a copy
of the ldapConfig.json called ldapConfig.json.good and restore
ldapConfig.json from the copy before login tests commence.  The
ldap configs also used "true" and "false" strings instead of
boolean values; strings instead of booleans is now obsolete.

Finally, testing uncovered some bugs in setMsalConfig discussed
in Xc-11941.  These should now be fixed.

Change-Id: I0f46c67c04ff69392d50bd9cd42760769c7a384f