has_require var checks should handle nwjs case
XD checking for condition of running node context by checking
if 'require' is defined; some functionality is different
in this context.
However, when running xcalar-gui in nwjs, 'require' will be defined,
but xcalar-gui needs to have the same functionality as running
in the browser like normal.
When running xcalar-gui in nwjs, there will be a global 'nw' object
exposed, so adding a check alongside 'require' being defined,
that 'nw' should be undefined.

Change-Id: I19c1c8d789b99ca9eede82661e9436dbf4518964