GUI-12213 - Improper comparison of number of nodes in gui installer Step 5 with NumNodes in the license
This change ensures that NumNodes and NumUsers in the Xcalar license are converted to integers.  It also
ensures that the number of nodes entered in Step 5A is also converted to an integer.  Taken together,
this properly compares the number of hosts to be installed with the maximum number of hosts in the license.

Change-Id: I141f5ee0fa0e35724b5e82ad61bd5f2c9d6da9d4