GUI-12218 Allow login.html to login directly with ssoToken
This commit introduces 3 changes:
1) login.html now recognizes ssoToken. If ssoToken is detected, it
verifies with expServer and if ssoToken is valid, we by-pass the usual
login screen and the user is logged in directly.

2) expServer login/verifyToken endpoint added -- The endpoint probably
will survive, but the token involved is very likely to change to a
proper token, say JWT. This endpoint is both used to verify that the
token is valid, as well as provide userInfo associated with the token.

3) expServer login/with/HttpAuth endpoint added -- Again, this endpoint
probably will survive, but the token involved is very likely to change.
There could be multiple ways to authenticate with expServer, and they
all result in expServer granting a token on successful authentication.
Currently, HttpAuth takes in the following JSON argument in a POST:

{ "credentials": "base64-encoding-of-username:password" }

Currently, the token is really just a base64-encoding of the entire
userInfo, so that we can recover userInfo given token by base64-decoding
the token. We mitigate against tampering by computing a SHA-256 HMAC on
userInfo with a secret salt. We mitigate against replay attack with a 5
min expiration of the token.

Change-Id: I5da82f9173de9fb7b6ecd0c12f2ab8d651776700