GUI-12513 - Azure B2C authentication fails during oauth token validation
This change saves the value of msal.b2cEnabled to be saved into the
auth.js scope when msalConfig.json is loaded or saved.  Once there,
it is used to determine which authnetication key endpoint is used to
to validate Azure Oauth tokens.  It also adds two minutes of tolerance
when evaluating timestamps on Azure Outh tokens.

Change-Id: I4f10b29ac885053987d78318a5c55176e3001b8d