GUI-12998 - LDAP authentication broken where ldapConfigEnabled is false
This change adds missing return statements that cause the load of LDAP
configuration to stop if ldapConfigEnabled is set to false, the server
key file is not set, or the server key file does not exist.  It also
relocates the statement where ldapConfigSetup is set to true to a point
where that setup is fully validated.

This fixes problems disabled or incomplete LDAP configurations were
loaded and actually used, with problematic effects on the login page.

This change needs to be added to 1.4.1.

Change-Id: If1db25d0ca968b9e579da8a6298fb270638409c1