GUI-13036,13049 - Xcalar token authentication failing because browser/XD is not sending connect.sid cookie
The browser was not sending the cookie after the initial timeout
period expired because the expServer was not sending the cookie
back with expiration time updates with every subsequent call.

This made the session expiration times on the client and server
to fall out of sync to the point that the cookie in the browser
disappears.  Once that happens, the browser sends a request without
a cookie and the server treats it like a new, unauthenticated
session.

This change fixes that problem by changing the express-session
package rolling option to true.  The middleware layer now sends
and updated session cookie with each http response.

The change also addresses GUI-13049, which requires that
/login/msalConfig/get be call-able without authentication
checks because it is invoked on the login page.

Change-Id: I45135d04c70fcda53c8f436109d51e51c0f2fe6a