GUI-13143 - Thrift proxy requests to the expServer return 404 not 401 when not part of an authenticated session
By default, the express-http-proxy filter code returns status 404 with
the following message when an authenticated session is used:

< HTTP/1.1 404 Not Found
<!DOCTYPE html>
<html lang="en">
<head>
<meta charset="utf-8">
<title>Error</title>
</head>
<body>
<pre>Cannot POST /thrift/service/thrift/service/thrift/service/XcalarApiService/</pre>
</body>
</html>

This change alters the http status and the message body to this:

< HTTP/1.1 401 Unauthorized
Unauthorized thrift request

It does this by explictly making the filter failure return 404 and
"Unauthorized thrift request" with res.status().send().

Doing this requires reordering the layers of express middleware.
The express-http-proxy code must come before use of the bodyparser.
Otherwise, the bodyparser attempts to parse the bodies of thrift
requests.  The proxy code must now come after the Access-Control-Allow
header modification requests, because attempting to alter these header
fields after a proxy failure causes errors to occur.

Change-Id: I9292c2b38e61b46fc767165782d660d4e275f1f4