CP1 and later - Jupyter proxy for the expServer
This change creates a second http server port for the expServer to listen
on for the exclusive purpose of proxying Jupyter traffic to enable only
authenticated access to Jupyter.

The second port was found to be necessary because Jupyter uses both
http traffic and websocket traffic.  An additional proxy for the http
traffic is easy to add to the express middleware stack.  Handling
Jupyter websocket traffic is more complex. The expServer already uses
Socket.io to handle client/server communication and Socket.io uses
websocket.  The Jupyter websocket traffic needs to be proxied, not
handled.  Attempts to identify websocket upgrade requests and proxy
them for Jupyter, results in proxying both Jupyter and Socket.io
traffic to Jupyter.  Tests show that Socket.io code does not recognize
the Jupyter websocket requests for processing (which causes the Jupyter
requests to be completely dropped.)

Separating the jupyter traffic to a second port (at second port number
1 higher than the first) allows a second express middleware app to just
have the http proxy and a dedicated websocket proxy attached solely
for the use of Jupyter that does not interact with the Xcalar Socket.io
traffic.

The environment variables XCE_THRIFT_PORT and XCE_JUPYTER_PORT allow
the target ports for traffic forwarding to be changed without modifying
the expServer.

Change-Id: I7413964ad667d1f5e67c65303c094fff5c25e5af