GUI-13350 - CP1 and later - expServer jupyter proxy port number is being treated as string in gui-installer
This change fixes a problem where port numbers taken from environment
variables are treated as strings, not numbers.  The strings were implicitly
converted to numbers when the single http listener port was used.
When a second port for jupyter proxying was added, which is supposed
to be (port + 1), this was treated as string concatenation instead of
addition. This resulted in an implicit port number of 122241 (not 12225) for
the gui-installer, which is out of range.

Change-Id: I8ad035b95ab416c7b63ba9867fd2fbc0ed6fbf5b