GUI-13416 - SQL history might be too large for expServer
The express-http-proxy limits the size of the body of a proxied http
request to be 1MB in size by default. This is too small for Xcalar
thrift requests (including one SQL history request that was over 1MB
in size in the bug). This change alters the size limit for the thrift
and jupyter proxies to the 20MB used for other expServer REST calls.
It also configures the thrift proxy to not parse the thrift requests.

CP1 and later

Change-Id: I3d2a2fef7865d8663a9faa302478e24408391af1