GUI-13416 CP1 Failed to commit kvStore meta info as api calls failed at expServer (CP1 and later)
The symptom motivating this bug is the appearance of ECONNRESET error messages
in the expServer log for proxied thrift calls:

2018-10-04T18:25:29.066000Z-0700:Xcalar ExpServer ERROR: ECONNRESET error on proxy { Error: socket hang up
    at createHangUpError (_http_client.js:258:15)
    at Socket.socketCloseListener (_http_client.js:290:23)
    at emitOne (events.js:101:20)
    at Socket.emit (events.js:188:7)
    at TCP._handle.close [as _onclose] (net.js:509:12) code: 'ECONNRESET' }

The cause for these TCP connection resets was traced to the parseReqBody: false
configuration option in the thift proxy express middleware object.

The exact reason why setting the value to false causes the resets is unknown.
The false setting does significantly change the network behavior of the proxy,
however.  When false, the proxied data is sent in multiple TCP frames that
need to be reassembled to create the complete request.  When true, each thrift
request is sent as a single frame, with a properly set Content-Length.

Something about the network changes when set to false causes Mgmtd to silently
close connections, or the express middleware to claim that it is.

This change does two things: it sets parseReqBody explicitly to true (with
notation,) and it causes ECONNRESET errors to be explicity returned as http 502
errors.  If ECONNRESET errors are not explicity handled, they are ignored by
the proxy code according to the documentation.

Change-Id: I16e75f965d9e993b3740483b4b67809c8d531757