GUI-13557 - CP1 and later - Could not get logs: admin GUI screen
It was not possible to monitor logs from the XD admin screen.  This
is due to the fact that the call to run the Master/Slave command to
collect the logs was not updated to use session cookies for
communication.  The slave requests to collect logs returned a 401
Not Authorized because no session cookie to authenticate was
included in the request header.

This change fixes broken function call.

Change-Id: I027aec60123509802c82fcd4fcde16cd5c1752f6