Xc-13508 - CP1 and later - "email" specified in defaultAdmin.json dictates which user name the workbooks appear under
Though reported for default admin, this applies to all cases where
username is different from the e-mail address (like Active Directory
LDAP.)

The root cause of the problem was introduced in SHA 5ef18d13 where the session username
was changed from the browser session storage xcalar-username:

       const username: string = xcSessionStorage.getItem("xcalar-username");
       const user: XcUser = new this(username, stripEmail, gCollab);

to the e-mail address returned from the session token:

       let setCurrentUserHelper: Function = (username, isAdmin) => {
            const user: XcUser = new this(username, isAdmin);

       XcUser.checkCurrentUser()
        .then((data) => {
            setCurrentUserHelper(data.emailAddress, data.admin);
            deferred.resolve();
        })

This works for OpenLDAP (where username and e-mail address are identical) but breaks
for default admin and Active Directory (where they are not.)

This change adds the xiusername passed to the expServer to the session token as the
username field.  This field is retrieved by /auth/sessionStatus and the
setCurrentUserHelper is modified to use the retrieved username instead of the
emailAddress.

Change-Id: I77d056fe28dce7d5df4373eead230b13811ddf21