GUI-13938 - expServer search of default.cfg cannot tolerate commented out definitions of XcalarRootCompletePath that precede the actual one.
This change replaces code that searches for the presence of XcalarRootCompletePath
with a regular expression that better exclude commented lines and deal with spaces.

The previous code had trouble with cases like this:

// Constants.XcalarRootCompletePath=/some/path/that/does/not/exist
Constants.XcalarRootCompletePath=/var/opt/xcalar

It determined /some/path/that/does/not/exist to be the correct path.

The regex can now anchor Constants to the start of the line, excluding
the commented line(s.)

Change-Id: I94488572070d33c930fbdb2a8ad1153c8eaf6bf2