GUI-13948 CP1: Parameterize HTML branding of XPE Guis (part 1: Builld time)
Right now, the HTML is hard coding the name of the xpe app in its GUIs.
There are two ways branding occurs in these GUIs:
1. html (<title> tags, etc.).
2. js (status messages, err msgs, etc).

This change takes care of the branding which occurs in the HTML.
 Now the HTML files will reference a common struct in assets/lang/en/htmlTStr.js
 whenever the app name needs to be referenced.  Simply update that entry in htmlTStr.js
 and all the html files will get the update.

Am waiting on 2. because I need to think of the best way to handle it for the app.

Change-Id: I43d25e88b0171d9a93f94325013e3458bde2dc67