GUI-14038 - XD logs in and immediately logs out for LDAP logins
The login failure in the bug is due to the fact that the LDAP user
in question did not have an email address in Active Directory.  This
was acceptable for LDAP data collection during login, but it created
a session token without a mail field.  This failed session token
status checking (/auth/sessionStatus) and returned a null token.
In response, the XD interface immediately logged the user out.

This change makes the first name and email fields optional by setting
returned values to empty strings if the data structure returned by
LDAP query does not contain the mail or cn fields.

Change-Id: I4d9153906d5c5db4ea9692e8101bc5701342b095