Simple login/logout authentication against vault ldap
This change adds methods for authenticating Xcalar users against the
vault ldap interface.

It requires a configuration file called vaultConfig.json with permissions
600 in XcalarRootCompletePath/config.

The file should look like this:

{
  "vaultEnabled": true,
  "vault": {
    "vault_uri": "https://vault:8200",
    "vaultUserGroup": "xceusers",
    "vaultAdminGroup": "xceusers"
  }
}

The vaultUserGroup and vaultAdminGroup are the names of policies returned
by a successful auth.

The data returned by the vault login is stored in a session token
credential called vault.

There is no installer support or XD login configuration modal yet.

Change-Id: I8c8da2fc0904aadfa50ec6b09036d6a50824181d