JIRA-IA-21 - repo to get access to expServer and XD
In order for the expServer to get access to Xcalar javascript client
code found in the assets/js/xcrpc directory, this change alters the
Gruntfile to cause the code to be made into an npm package tarfile,
and then the package installed during the build from that tarfile into
the services/expServer/node_modules directory.

The intermediate xcalar npm module tar file is deleted from the
source directory before source is copied to the destination location.

Change-Id: I9d47672dcb21fee997018143a71070e71dec2c65