SDK-157 service manager infra to get kvstore to expServer for converting
This change utilizes the component/dag module of XD in expServer in order
to convert a query stored in KVS format into a query string that can be executed
by XcalarQuery.

Steps that are done by grunt but leaving the comment for informational purposes.

* make trunk
* cp -R xcalar-gui/xcalar-gui/assets/js/components/dag/*
  xcalar-gui/xcalar-gui/services/expServer/dagHelper/

This copies the generated .js files from XD into a new dagHelper
directory in expServer.

Change-Id: I533a630fc9527537990545ccd1a65853b8e00551