TOOLS-188 - Error msgs from tar when running GUI installer script
npm build testing is setting the modification times of some files to
the start of the Unix epoch (31-Dec-1969).  Though it has no effect
on the operation of Xcalar, these files cause tar to produce a lot of
noisy warning messages about "implausibly old timestamp."

This change uses find and touch in Gruntfile.js to alter the
modification times of the affected files.

Change-Id: I45935d793fc3f0f17357c50aec1e03912cd63dd0