XD-501 Dio: Join - Remove keepAllColumns flag when calling API
We set the flag to false when calling XIApi, so that it can be removed
directly when backend/api change is done.

For convenience, we still keep the flag in dataModel/dagInput, so that
we don't have to specify the columns one by one. In a future change, the
flag will be removed from UI and can only be manipulated in adv.
form/json.

Change-Id: I2fac9a7a0cfcdf8aa3cb46b53319fb823c1b5666