Remove login page inline from Grunt build for non-installer builds
Turns out a bunch of our testing infrastructure does not ensure that
XD submodules are populated during the build.  The assets_inline
infrastructure requires that the XD submodule
3rd/microsoft-microsoft-authentication-library-for-js must be present
for inline in order for build to succeed.

Until that module can be guaranteed to be present, this change makes
the inline operation installer build only.

Change-Id: Ie6be8d6a392d8b4cc8c70fd532b1191871b55eac