SDK--304 Ship listXdfs output to expServer for KVS to query conversion
This provides the server side of the change to receive the listXdfs output
with the kvstore to query string conversion.  expServer doesn't use the thrift
interfaces and we don't want to introduce that as we're moving away from them.
When the xcrpc interfaces are available we can switch to using them.

Change-Id: I337ed82f0a91023ecc4fc9131b50b5db5b86fd3c