SDK-291 expServer consolidate KVS into query
If one or more dataflows in a workbook contain linkout/linkin nodes then we
can't just run such a dataflow in isolation.  So we try to run the specified
dataflow and if it returns an error (currently "DagTabManager is not defined")
then we'll try again except pass all the dataflows in the workbook.

Change-Id: I66d9ad7620060966225d9acb2b97b6bdf01c0645