Add xcalar-0.0.1 package to the expServer package.json file
The previous iteration of the build would modify the expServer package.json
file by running npm install to add the locally build xcalar package tgz file
during the build.  Here, we take the much safer approach of adding that dependency
to the file and modify the build process to support that dependency.

This will have the side effect of potentially causing an npm install error if
the xcalar tgz file is not built and present, but this can be fixed by running
a grunt build target that includes the init stage (trunk, installer, etc...)
just once.

npm install operations in the services/expServer directory are marked --no-save
because we do not want npm modifying the package.json/package-lock.json files
during the install/uninstall/install sequence.

Change-Id: I05db931063bf8d6debd50df6ec1b1721d0c73be6