TOOLS-393 - XD authentication fails when the username has uppercase letters
A change in Xcalar in June 2017 automatically caused all usernames entered
to be converted to lower case: https://gerrit.int.xcalar.com/#/c/5234

No one seems to remember exactly why this change was made; the strongest
possible reason was that given the state of login authentication at the
time (no sessions), the username taken directly from the login screen was
used to identify the Xcalar sessions and the desktop to use.  This meant
that different capitalization would cause the user's desktop to "disappear".

Therefore, the username was converted to lowercase as soon as it was read
out of the login page html.

Today, there is still a some need for canonical form in username (we still
identify sessions by username) but we have a session-token mechanism that
can and should more effectively resolve issues like this.

There is also a desire to enforce case sensitivity in actual authentication
(so Admin might work, but admin might not) since some authentcators (like
defaultAdmin) are case sensitive.

This change removes the conversion to lower case out of the login page,
and puts it in the routines that prepare session token metadata.  This
allows case sensitivity in authentication (for methods that support it)
but provide a canonical session name to indentify user workbooks.

Change-Id: I0868aeee7cb6f673c78b3bd2f15052efa19e69dd