TOOLS-390 - 502 Server Error: Bad Gateway during RC15 Release Test
The underlying bug deals with highly loaded release test deployments
where the server is not responding in a timely manner.  To deal with
this problem, changes in https://gerrit.int.xcalar.com/#/c/15871/
propose to add re-try mechanisms to SDK authentication code.

These changes are stateful -- they add a cookie cache much like what
a browser uses to SDK code.  This means that http(s) requests now
inherit cookies, including session cookies, from requests in an
SDK application's prior history.

This interferes with agreed upon behavior of the /login and
/auth/serviceSession -- providing a session cookie to these
endpoints causes the existing to be modified, and not a new one
created.

This change deals with the problem by adding middleware to the
expServer software stack that filters out session cookies to
the /login and /auth/serviceSession calls.  Without a provided
cookie, a new, different one is automatically generated.

Change-Id: I31701b544f310bc8f1202dfa57332b4376a2710d