SDK-757 - SDK-687 fails to take expServer use in gui-installer into account
The expServer is used in two cases: during normal operation and in the gui-installer.
Changes for SDK-687 add a mandatory search for the number of the current node. This
is fine for normal operation, but does not apply in the gui-installer case and
causes the installer to fail.

This changes uses presence of XCE_INSTALLER_ROOT and the lack of XCE_CONFIG to
detect that the expServer is running inside an installer, and therefore can just
be node 0 for purposes of the expServer.

Change-Id: Ie6467c4a472eefb0feae279969c611670c281a78