XD-1985 Fix a potential bug to use inner variable in DagNodeMenu.ts
The original code in DagaNodeMenu.ts use $dfWrap and $menu and assume they
are alreaday declared, but if may cause potential crash if function is called before
the variable initialization is done. The crash occurs in Daily Test.

To fix this, change to use function call to get the element dynamically.

Change-Id: I61b871c129e774f9d10a422eb24235ca57aa9c5d