SDK-793 Create symlink to lowest usrnode ID UDS
In addition to UDS /tmp/xcalar_sock/usrnode-<NodeId>, we will create
symlink /tmp/xcalar_sock/usrnode.

For the multi-usrnode on the same host case, the usrnode with the lowest
node ID get's to establish the symlink to /tmp/xcalar_sock/usrnode.

Also delete the userNode detection code in the expServer

Change-Id: Id24814f8c2f4e7b6ffadd21fb9a5773e5193e317