SDK-770 SQL/XD node name catch22
This change is to support querying xcalar tables which contains # in it.
For example: There is a xcalar table with name t1#123, query should be like
"select * from `t1#123`".

To make this work, sql parser is changed to modifiy the sql by replacing # with _ with
this change https://gerrit.int.xcalar.com/#/c/16963/

sql parser returns a modified sql and modified identifiers. This is the new output of sql parser
[{"newIdentifiers":{"`t1_123`":"`t1#123`"}, "newSql":"select * from `t1_123`", "identifiers":["`t1_123`"], "sql":"select * from `t#123`"}]

The code changes in this commit will make use of the modified sql/identifiers to properly do
send schema and get logical plan.

Change-Id: Ia8eb03335243dbd0ea9b639d55ca6c5f74b9fefe